requirements.txt
dataclasses
emoji>=1.6.1
python-telegram-bot>=13.13
validators
